EP26 BugFix - Deleting results with missing files.
Issue with deleting submissions if related files have already been deleted and system is reloaded.
Caused by report object having to be re-generated on a system reload which cannot be done once a file no longer exists.
Decided if a file is deleted we don't want the reports as they are not as useful and would require larger rewrite.
Deleting results when a submission is deleted prevents this issue from occurring.